+++ CHANGELOG.md
@@ -12,6 +12,9 @@ speaking, considered to be the player's experience.
 
 ### Added
 
+- Added a clone button to Caravans.  If there is an idle Caravan, it is used as
+  the clone.  Otherwise, a new Caravan spawns at the Market closest to the
+  Caravan's first stop.
 - Show the skill icon of the unit being trained in Towns under the progress bar
   on the world map and in the info pane.